FreeNAS has an NFSv4.1 server but it doesn't work with ESXi 6.0
From the packet captures ESXi is violating the NFS RFC

VMWare incorrectly sets the field in reclaimcomplete to true.
The RFC pretty clearly states that the "global" one that is done at
the end of reclaiming locks should be "false".

(It is easy to hack the server code to just ignore the "false". You
could try that and see what happens.)

You can look at RFC5661 (around page #567) and I think it is pretty
clear that the argument should be false.